pyproject.toml
[tool.poetry]
name = "dungeon_lad"
version = "0.1.0"
description = ""
authors = ["Skyler Burger <skylerburger@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
python-environ = "^0.4.54"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
